requirements.txt
# requirements.txt
flask==2.0.1
flask-pymongo==2.3.0
flask-cors==3.0.10
pymongo==4.3.3
python-dotenv==0.19.2
marshmallow==3.14.1
Flask-Bcrypt==1.0.1
PyJWT==2.3.0
Werkzeug==2.2.3
faker
